Move history return types to optimagic.typing
