+++ src/optimagic/typing.py
@@ -2,6 +2,9 @@ from dataclasses import dataclass, fields
 from enum import Enum
 from typing import Any, Callable, ItemsView, Iterator, KeysView, Literal, ValuesView
 
+import numpy as np
+from numpy._typing import NDArray
+
 PyTree = Any
 PyTreeRegistry = dict[type | str, dict[str, Callable[[Any], Any]]]
 Scalar = Any
@@ -66,3 +69,35 @@ class OptimizationType(str, Enum):
 
 
 OptimizationTypeLiteral = Literal["minimize", "maximize"]
+
+
+@dataclass(frozen=True)
+class IterationHistory(DictLikeAccess):
+    """History of iterations in a process.
+
+    Attributes:
+        params: A list of parameters used in each iteration.
+        criterion: A list of criterion values obtained in each iteration.
+        runtime: A list or array of runtimes associated with each iteration.
+
+    """
+
+    params: list[PyTree]
+    criterion: list[float]
+    runtime: list[float] | NDArray[np.float64]
+
+
+@dataclass(frozen=True)
+class MultiStartIterationHistory(TupleLikeAccess):
+    """History of multiple start iterations.
+
+    Attributes:
+        history: The main iteration history, representing the best end value.
+        local_histories: Optional, a list of local iteration histories.
+        exploration: Optional, iteration history for exploration steps.
+
+    """
+
+    history: IterationHistory
+    local_histories: list[IterationHistory] | None = None
+    exploration: IterationHistory | None = None
